Matrix calculus: Move summation outside of the case
The derivative is a sum of N vectors, some of which are zero and some of
which are not. Moving the summation inside the piecewise curly brace
makes this unclear. The conditional is decided separately for each i,
and i is only defined inside the summation, so it doesn't really make
sense to move the summation inside the curly brace.